Refactor login process
Now a controller wants to log a user in, it can redirect them to
`/login` and the login controller will do the OIDC login and
return the user to the homepage.

If a `?returnUri` URL parameter is set on the redirect to `/login`
eg. `/login?returnUri=http://localhost:3000/vouch` then once the
user is returned from OIDC login it will redirect them back to
that URI.

This is mostly stolen from the other Solid proof of concept app.