Save a fake photo to the app's storage
In the real journey a user would upload a photo of themselves.
Instead, fetch a fake image from thispersondoesnotexist.com and
store that locally in the app and write its path to the user's
session.

This allows us to easily display the image on the next page
which will be the confirmation page (since it'll have a public URL)
at which point we can put it in the user's pod