Use Universal Access to give the voucher access controls
This isn't the way I'd ideally want to do it, but expedient for the
hack. This grants the voucher access controls to the container we've
just made to save a vouch-request-vc into.

Theoretically that gives the user write access over the request vc which
we don't want.

With a bit more time I'd ideally issue two access request modifications:

- Read access to the request-vc
- Write access to a container where the Vouch can go.

We'd leave it to an app the user can log into to clear up old requests
if we needed that (the model being clean up can happen from an app the
user logs into and trusts, acting on their own pod).

What stopped us doing that here was giving ACP permissions on a
container that didn't exist. After a brief attempt to create a container
and THEN do the ACP, we reverted to this simpler solution to save some
debugging time at the hackathon.