Remove the 2 year requirement for vouching
In this prototype, we want to explore the idea that anyone can vouch for
anyone else, and that different vouches might have different weights.

For example, if I vouched for someone that I'd only known for a few
months, that would be potentially less valuable than if I vouched for
someone I'd known for years.

It also opens up the scope to attach professional credentials to the
strength of a vouch too - if a doctor vouches for someone that's
probably of higher strength than a random citizen.